chore(iOS): add previews and disrupted branch test
